Removes DOM class and merges code into AsyncCSS.
Why?

Want to keep it simple in this version to avoid having to change the DOM class later. It's not needed right now.
